Created a README
Having a readme is an important thing to have in your project.  I haven't filled in the details yet, but I will eventually.  And I'll add a bunch of details about what this project is, how to run it, why I created it, etc.  It will help many users in the future pick up where I left off.

This is actually a test of multiline git commit messages.